[eosioclient] moved into library
