+++ libraries/CMakeLists.txt
@@ -7,3 +7,4 @@ add_subdirectory( appbase )
 add_subdirectory( chain )
 add_subdirectory( testing )
 add_subdirectory( abi_generator )
+add_subdirectory( eosioclient )
