+++ programs/eosioc/CMakeLists.txt
@@ -1,5 +1,3 @@
-add_subdirectory(eosioclient)
-
 add_executable( eosioc main.cpp httpc.cpp help_text.cpp localize.hpp config.hpp CLI11.hpp)
 if( UNIX AND NOT APPLE )
   set(rt_library rt )
